Rename controller file
